+++ SistemaIndicadoresAmbientales/Entity/ConsumoEntradaActualizar.cs
@@ -4,7 +4,7 @@ namespace SistemaIndicadoresAmbientales.Entity
 {
     public class ConsumoEntradaActualizar
     {
-        public int Cantidad { get; set; }
+        public float Cantidad { get; set; }
         public string Fecha { get; set; }
         public int Id_Consumo_Compost_Entrada { get; set; }
     }
