+++ SistemaIndicadoresAmbientales/Entity/HistoricoElectrico.cs
@@ -5,7 +5,7 @@
         public int Id_Consumo_Electrico { get; set; }
         public int Anio { get; set; }
         public int Numero_Vatihorimetro { get; set; }
-        public float Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public int Mes { get; set; }
     }//end class
